requirements.txt
torch
torchvision
torchaudio
timm
opencv-python
gradio
scikit-learn
matplotlib
tqdm
kaggle
Pillow
numpy